requirements.txt
toolz
colorama
ryanair-py
streamlit